requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
geopandas>=0.14.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0 